EUW-486: Rename frå issuer-ui-demo til issuer-ui
