+++ catalog-info.yaml
@@ -1,15 +1,15 @@
 apiVersion: backstage.io/v1alpha1
 kind: Component
 metadata:
-  name: digital-lommebok-issuer-ui-demo
-  description: EUDIWs issuer-ui-demo
+  name: digital-lommebok-issuer-ui
+  description: EUDIWs issuer-ui
   annotations:
-    github.com/project-slug: felleslosninger/eudiw-issuer-ui-demo
+    github.com/project-slug: felleslosninger/eudiw-issuer-ui
 spec:
   type: application
   lifecycle: test
   owner: team-digital-lommebok
-  system: digital-lommebok-issuer-ui-demo
+  system: digital-lommebok-issuer-ui
   dependsOn:
     - component:digital-lommebok-issuer-server
