+++ docker-compose.yaml
@@ -5,7 +5,7 @@ networks:
 
 services:
 
-  issuer-ui-demo:
+  issuer-ui:
     build:
       context: .
       dockerfile: docker/dev.Dockerfile
@@ -16,17 +16,17 @@ services:
       spring_profiles_active: docker
       _JAVA_OPTIONS: "-XX:+IgnoreUnrecognizedVMOptions -XX:UseSVE=0 -Dotel.metrics.exporter=none -Dotel.traces.exporter=none -Dotel.span.exporter=none -Dotel.logs.exporter=none"
     ports:
-      - "9290:9290"
-      - "9295:9295"
+      - "9292:9292"
+      - "9293:9293"
     networks:
       - digital-lommebok
     entrypoint: "/bin/bash -c"
     command:
       - |
         cd /usr/local/tomcat/webapps/
-        java -javaagent:opentelemetry-javaagent.jar -agentlib:jdwp=transport=dt_socket,server=y,suspend=n,address=*:9295 -jar application.jar
+        java -javaagent:opentelemetry-javaagent.jar -agentlib:jdwp=transport=dt_socket,server=y,suspend=n,address=*:9293 -jar application.jar
     healthcheck:
-      test: wget --no-verbose --tries=1 --spider http://localhost:9290/health/readiness || exit 1
+      test: wget --no-verbose --tries=1 --spider http://localhost:9292/health/readiness || exit 1
       interval: 15s
       start_period: 5s
       timeout: 5s
