+++ pom.xml
@@ -8,10 +8,10 @@
         <version>3.5.5</version>
     </parent>
     <groupId>no.idporten.eudiw</groupId>
-    <artifactId>eudiw-issuer-ui-demo</artifactId>
+    <artifactId>eudiw-issuer-ui</artifactId>
     <version>DEV-SNAPSHOT</version>
-    <name>eudiw-issuer-ui-demo</name>
-    <description>EUDIW issuer-ui for demo/testing purposes</description>
+    <name>eudiw-issuer-ui</name>
+    <description>EUDIW issuer-ui</description>
     <properties>
         <java.version>24</java.version>
         <maven-compiler-plugin.version>3.14.0</maven-compiler-plugin.version>
