+++ src/main/java/no/idporten/eudiw/issuer/ui/IssuerUiApplication.java
@@ -1,14 +1,14 @@
-package no.idporten.eudiw.issuer.ui.demo;
+package no.idporten.eudiw.issuer.ui;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 
 @SpringBootApplication
-public class DemoApplication {
+public class IssuerUiApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(DemoApplication.class, args);
+		SpringApplication.run(IssuerUiApplication.class, args);
 	}
 
 }
