--- src/main/java/no/idporten/eudiw/issuer/ui/demo/web/JsonRequest.java
@@ -1,4 +1,0 @@
-package no.idporten.eudiw.issuer.ui.demo.web;
-
-public record JsonRequest(String json) {
-}
