+++ src/main/java/no/idporten/eudiw/issuer/ui/exception/IssuerUiExceptionHandler.java
@@ -1,6 +1,6 @@
-package no.idporten.eudiw.issuer.ui.demo.exception;
+package no.idporten.eudiw.issuer.ui.exception;
 
-import no.idporten.eudiw.issuer.ui.demo.issuer.config.IssuerServerProperties;
+import no.idporten.eudiw.issuer.ui.issuer.config.IssuerServerProperties;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
