+++ src/main/java/no/idporten/eudiw/issuer/ui/issuer/IssuerServerService.java
@@ -1,9 +1,9 @@
-package no.idporten.eudiw.issuer.ui.demo.issuer;
+package no.idporten.eudiw.issuer.ui.issuer;
 
-import no.idporten.eudiw.issuer.ui.demo.exception.IssuerServerException;
-import no.idporten.eudiw.issuer.ui.demo.exception.IssuerUiException;
-import no.idporten.eudiw.issuer.ui.demo.issuer.config.IssuerServerProperties;
-import no.idporten.eudiw.issuer.ui.demo.issuer.domain.IssuanceResponse;
+import no.idporten.eudiw.issuer.ui.exception.IssuerServerException;
+import no.idporten.eudiw.issuer.ui.exception.IssuerUiException;
+import no.idporten.eudiw.issuer.ui.issuer.config.IssuerServerProperties;
+import no.idporten.eudiw.issuer.ui.issuer.domain.IssuanceResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
