+++ src/main/java/no/idporten/eudiw/issuer/ui/issuer/domain/TxCode.java
@@ -1,4 +1,4 @@
-package no.idporten.eudiw.issuer.ui.demo.issuer.domain;
+package no.idporten.eudiw.issuer.ui.issuer.domain;
 
 
 import com.fasterxml.jackson.annotation.JsonProperty;
