+++ src/main/java/no/idporten/eudiw/issuer/ui/web/IssuanceRequest.java
@@ -1,4 +1,4 @@
-package no.idporten.eudiw.issuer.ui.demo.web;
+package no.idporten.eudiw.issuer.ui.web;
 
 import java.io.Serializable;
