+++ src/main/java/no/idporten/eudiw/issuer/ui/web/JsonRequest.java
@@ -1,0 +1,4 @@
+package no.idporten.eudiw.issuer.ui.web;
+
+public record JsonRequest(String json) {
+}
